Generalize loader loading
The state of the AMD loader must be refreshed before each test. Although
the `context` configuration option serves this purpose for RequireJS, it
is not implemented by other script loaders. Using a dedicated iFrame as
the context for each test will allow for *all* AMD loaders to be
re-initialized before every test.